added modules to the project
